Move X.509-related helpers into a folder
