+++ packages/server/src/helpers/x509/convertX509BufferToPEM.ts
@@ -1,9 +1,9 @@
-import { base64url } from './index.ts';
+import { base64url } from '../index.ts';
 
 /**
  * Convert buffer to an OpenSSL-compatible PEM text format.
  */
-export function convertCertBufferToPEM(certBuffer: Uint8Array): string {
+export function convertX509BufferToPEM(certBuffer: Uint8Array): string {
   // We specifically do not want base64url used here, PEM is base64
   const b64cert = base64url.bufferToBase64(certBuffer);
