+++ packages/server/src/helpers/x509/mapX509SignatureAlgToCOSEAlg.ts
@@ -1,4 +1,4 @@
-import { COSEALG } from '../cose.ts';
+import { COSEALG } from '../../cose.ts';
 
 /**
  * Map X.509 signature algorithm OIDs to COSE algorithm IDs
